Abstract ASN1 structures into different files
