+++ Sources/_CryptoExtras/Util/EncryptedPEMDocument.swift
@@ -64,7 +64,7 @@ struct EncryptedPEMDocument: PEMRepresentable {
         switch algorithm {
         case .pkcs5PBES2:
             let pbes2Params = try PBES2Parameters(asn1Any: parameters)
-            let pbkdf2Params = try PBKDF2Parameters(asn1Any: pbes2Params.keyDerivationFunction.parameters)
+            let pbkdf2Params = try KeyDerivationFunction.PBKDF2Parameters(asn1Any: pbes2Params.keyDerivationFunction.parameters)
             
             let hashFunction = pbkdf2Params.hashFunction
             
@@ -128,174 +128,6 @@ extension ASN1ObjectIdentifier {
         case .aes256_CBC: 32
         case .des_EDE3_CBC: 24
         default: fatalError("Not an encryption algorithm")
-        }
-    }
-}
-
-extension KDF.Insecure.PBKDF2.HashFunction {
-    static func from(objectIdentifier: ASN1ObjectIdentifier) -> Self? {
-        switch objectIdentifier.oidComponents {
-        case [2, 16, 840, 1, 101, 3, 4, 2, 1],
-            [1, 2, 840, 113549, 2, 9]: // hmacWithSHA256
-            .sha256
-        case [2, 16, 840, 1, 101, 3, 4, 2, 2]:
-            .sha384
-        case [2, 16, 840, 1, 101, 3, 4, 2, 3]:
-            .sha512
-        default: nil
-        }
-    }
-}
-
-struct PBES2Parameters: DERImplicitlyTaggable {
-    static var defaultIdentifier: SwiftASN1.ASN1Identifier { .sequence }
-    
-    let keyDerivationFunction: KeyDerivationFunction
-    let encryptionScheme: EncryptionScheme
-    
-    init(keyDerivationFunction: KeyDerivationFunction, encryptionScheme: EncryptionScheme) {
-        self.keyDerivationFunction = keyDerivationFunction
-        self.encryptionScheme = encryptionScheme
-    }
-    
-    init(derEncoded: SwiftASN1.ASN1Node, withIdentifier identifier: SwiftASN1.ASN1Identifier) throws {
-        self = try DER.sequence(derEncoded, identifier: identifier) { nodes in
-            let keyDerivationFunction = try KeyDerivationFunction(derEncoded: &nodes)
-            let encryptionScheme = try EncryptionScheme(derEncoded: &nodes)
-            
-            return .init(keyDerivationFunction: keyDerivationFunction, encryptionScheme: encryptionScheme)
-        }
-    }
-    
-    func serialize(into coder: inout SwiftASN1.DER.Serializer, withIdentifier identifier: SwiftASN1.ASN1Identifier) throws {
-        try coder.appendConstructedNode(identifier: identifier) { coder in
-            try self.keyDerivationFunction.serialize(into: &coder)
-            try self.encryptionScheme.serialize(into: &coder)
-        }
-    }
-}
-
-struct KeyDerivationFunction: DERImplicitlyTaggable {
-    static var defaultIdentifier: ASN1Identifier { .sequence }
-    
-    let algorithm: ASN1ObjectIdentifier
-    let parameters: ASN1Any
-    
-    init(algorithm: ASN1ObjectIdentifier, parameters: ASN1Any) {
-        self.algorithm = algorithm
-        self.parameters = parameters
-    }
-    
-    init(derEncoded: ASN1Node, withIdentifier identifier: ASN1Identifier) throws {
-        self = try DER.sequence(derEncoded, identifier: identifier) { nodes in
-            let algorithm = try ASN1ObjectIdentifier(derEncoded: &nodes)
-            let parameters = try ASN1Any(derEncoded: &nodes)
-            
-            return .init(algorithm: algorithm, parameters: parameters)
-        }
-    }
-    
-    func serialize(into coder: inout DER.Serializer, withIdentifier identifier: ASN1Identifier) throws {
-        try coder.appendConstructedNode(identifier: identifier) { coder in
-            try self.algorithm.serialize(into: &coder)
-            try self.parameters.serialize(into: &coder)
-        }
-    }
-}
-
-struct EncryptionScheme: DERImplicitlyTaggable {
-    static var defaultIdentifier: SwiftASN1.ASN1Identifier { .sequence }
-    
-    let encryptionAlgorithm: ASN1ObjectIdentifier
-    let encryptionAlgorithmParameters: ASN1OctetString
-    
-    init(encryptionAlgorithm: ASN1ObjectIdentifier, encryptionAlgorithmParameters: ASN1OctetString) {
-        self.encryptionAlgorithm = encryptionAlgorithm
-        self.encryptionAlgorithmParameters = encryptionAlgorithmParameters
-    }
-    
-    init(derEncoded: SwiftASN1.ASN1Node, withIdentifier identifier: SwiftASN1.ASN1Identifier) throws {
-        self = try DER.sequence(derEncoded, identifier: identifier) { nodes in
-            let encryptionAlgorithm = try ASN1ObjectIdentifier(derEncoded: &nodes)
-            let encryptionAlgorithmParameters = try ASN1OctetString(derEncoded: &nodes)
-            
-            return .init(encryptionAlgorithm: encryptionAlgorithm, encryptionAlgorithmParameters: encryptionAlgorithmParameters)
-        }
-    }
-    
-    func serialize(into coder: inout SwiftASN1.DER.Serializer, withIdentifier identifier: SwiftASN1.ASN1Identifier) throws {
-        try coder.appendConstructedNode(identifier: identifier) { coder in
-            try self.encryptionAlgorithm.serialize(into: &coder)
-            try self.encryptionAlgorithmParameters.serialize(into: &coder)
-        }
-    }
-}
-
-//  PBKDF2-params ::= SEQUENCE {
-//    salt CHOICE {
-//      specified       OCTET STRING,
-//      otherSource     AlgorithmIdentifier {{PBKDF2-SaltSources}}
-//    },
-//    iterationCount    INTEGER (1..MAX),
-//    keyLength         INTEGER (1..MAX) OPTIONAL,
-//    prf               AlgorithmIdentifier {{PBKDF2-PRFs}} DEFAULT algid-hmacWithSHA1
-//  }
-struct PBKDF2Parameters: DERImplicitlyTaggable {
-    struct HashFunction: DERImplicitlyTaggable {
-        static var defaultIdentifier: ASN1Identifier { .sequence }
-        
-        let objectIdentifer: ASN1ObjectIdentifier
-        let null: ASN1Null
-        
-        init(objectIdentifer: ASN1ObjectIdentifier, null: ASN1Null) {
-            self.objectIdentifer = objectIdentifer
-            self.null = null
-        }
-        
-        init(derEncoded: ASN1Node, withIdentifier identifier: ASN1Identifier) throws {
-            self = try DER.sequence(derEncoded, identifier: identifier) { nodes in
-                let objectIdentifer = try ASN1ObjectIdentifier(derEncoded: &nodes)
-                let null = try ASN1Null(derEncoded: &nodes)
-                
-                return .init(objectIdentifer: objectIdentifer, null: null)
-            }
-        }
-        
-        func serialize(into coder: inout DER.Serializer, withIdentifier identifier: ASN1Identifier) throws {
-            try coder.appendConstructedNode(identifier: identifier) { coder in
-                try self.objectIdentifer.serialize(into: &coder)
-                try self.null.serialize(into: &coder)
-            }
-        }
-    }
-        
-    static var defaultIdentifier: ASN1Identifier { .sequence }
-    
-    let salt: ASN1OctetString
-    let iterationCount: any ASN1IntegerRepresentable
-    let hashFunction: HashFunction
-    
-    init(salt: ASN1OctetString, iterationCount: any ASN1IntegerRepresentable, hashFunction: HashFunction) {
-        self.salt = salt
-        self.iterationCount = iterationCount
-        self.hashFunction = hashFunction
-    }
-    
-    init(derEncoded: ASN1Node, withIdentifier identifier: ASN1Identifier) throws {
-        self = try DER.sequence(derEncoded, identifier: identifier) { nodes in
-            let salt = try ASN1OctetString(derEncoded: &nodes)
-            let iterationCount = try Int(derEncoded: &nodes)
-            let hashFunction = try HashFunction(derEncoded: &nodes)
-            
-            return .init(salt: salt, iterationCount: iterationCount, hashFunction: hashFunction)
-        }
-    }
-    
-    func serialize(into coder: inout DER.Serializer, withIdentifier identifier: ASN1Identifier) throws {
-        try coder.appendConstructedNode(identifier: identifier) { coder in
-            try self.salt.serialize(into: &coder)
-            try self.iterationCount.serialize(into: &coder)
-            try self.hashFunction.serialize(into: &coder)
         }
     }
 }
